CC Wall thickness update
there is now a variable for combustion chamber wall thickness. This also thickens the walls around the throat. Printing small (r=1.5) throats has been a problem because the short layer times don't allow for adequate cooling and the slicer adds a lot of extra moves that decrease surface quality, so the extra thickness might help with printing cleaner throats.